+++ internal/useCase/getStorageChannelList/structure.go
@@ -1,13 +1,12 @@
 package getStorageChannelList
 
-type ChannelList []Channel
+type ResponseChannelList []ResponseChannel
 
-type Channel struct {
+type ResponseChannel struct {
 	Id          int    `json:"id"`
-	ChannelId   int    `json:"channel_id"`
+	OuterId     int    `json:"channel_id"`
 	Username    string `json:"username"`
 	Description string `json:"description"`
 	MemberCount int    `json:"member_count"`
-	CreatedAt   string `json:"created_at"`
 	UpdatedAt   string `json:"updated_at"`
 }